clean up and deduplicate webdav DELETE code
